+++ src/angular/planit/src/app/assessment/risk-popover/risk-popover.component.ts
@@ -3,6 +3,7 @@ import { Component, Input, OnInit, ViewChild } from '@angular/core';
 import { PopoverDirective } from 'ngx-bootstrap/popover';
 
 import {
+  City as ApiCity,
   Indicator,
   IndicatorService
 } from 'climate-change-components';
@@ -12,6 +13,7 @@ import { ModalTemplateComponent } from '../../shared/modal-template/modal-templa
 import { Location } from '../../shared/models/location.model';
 import { Risk } from '../../shared/models/risk.model';
 
+import { CityService } from '../../core/services/city.service';
 import { UserService } from '../../core/services/user.service';
 
 @Component({
@@ -23,7 +25,7 @@ export class RiskPopoverComponent implements OnInit {
 
   public indicators: Indicator[];
   public selectedIndicator: Indicator;
-  public city: Location;
+  public apiCity: ApiCity;
 
   @ViewChild('indicatorModal')
   private indicatorModal: ModalTemplateComponent;
@@ -32,13 +34,19 @@ export class RiskPopoverComponent implements OnInit {
   private popoverElement: PopoverDirective;
 
   constructor (private indicatorService: IndicatorService,
+               private cityService: CityService,
                private userService: UserService) {}
 
   ngOnInit() {
     this.updateRiskIndicators();
-    this.userService.current().subscribe(user => {
-      this.city = user.primary_organization.location;
-    });
+    this.userService.current()
+      .switchMap((user) => {
+        const id = user.primary_organization.location.properties.api_city_id;
+        return this.cityService.get(id);
+      })
+      .subscribe((apiCity) => {
+        this.apiCity = apiCity;
+      });
   }
 
   public openIndicatorModal(indicator: Indicator) {