+++ src/angular/planit/src/app/marketing/marketing.component.ts
@@ -1,12 +1,20 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
 
+import { AlertModule } from 'ngx-bootstrap';
 
 @Component({
   selector: 'app-marketing',
   templateUrl: './marketing.component.html',
   styleUrls: []
 })
-export class MarketingComponent {
+export class MarketingComponent implements OnInit {
 
-  constructor() {}
+  public showDownloadMessage = false;
+
+  constructor(private route: ActivatedRoute) {}
+
+  ngOnInit() {
+    this.route.queryParams.subscribe(params => this.showDownloadMessage = params.expired);
+  }
 }