+++ src/angular/planit/src/app/risk-wizard/steps/hazard-step/hazard-step.component.ts
@@ -5,6 +5,7 @@ import { Router } from '@angular/router';
 import { Subscription } from 'rxjs/Subscription';
 
 import {
+  City as ApiCity,
   Indicator,
   IndicatorService
 } from 'climate-change-components';
@@ -29,6 +30,7 @@ import { ModalTemplateComponent } from '../../../shared/modal-template/modal-tem
 import { RiskStepKey } from '../../risk-step-key';
 import { RiskWizardStepComponent } from '../../risk-wizard-step.component';
 
+import { CityService } from '../../../core/services/city.service';
 import { UserService } from '../../../core/services/user.service';
 
 interface HazardStepFormModel {
@@ -56,7 +58,7 @@ export class HazardStepComponent extends RiskWizardStepComponent<HazardStepFormM
   // Can't *ngFor a map type or iterable, so instead we realize the iterable and use that in *ngFors
   public directionalFrequencyOptionsKeys = Array.from(OrgRiskDirectionalFrequencyOptions.keys());
   public directionalIntensityOptionsKeys = Array.from(OrgRiskDirectionalIntensityOptions.keys());
-  public city: Location;
+  public apiCity: ApiCity;
   public indicators: Indicator[] = [];
 
   @ViewChild('indicatorChartModal')
@@ -73,6 +75,7 @@ export class HazardStepComponent extends RiskWizardStepComponent<HazardStepFormM
               protected toastr: ToastrService,
               protected router: Router,
               protected previousRouteGuard: PreviousRouteGuard,
+              private cityService: CityService,
               private userService: UserService,
               private fb: FormBuilder,
               private indicatorService: IndicatorService) {
@@ -85,9 +88,14 @@ export class HazardStepComponent extends RiskWizardStepComponent<HazardStepFormM
     this.setupForm(this.fromModel(this.risk));
     // Load initial risk indicators and subscribe to watch for weather event changes after
     this.updateRiskIndicators();
-    this.userService.current().subscribe(user => {
-      this.city = user.primary_organization.location;
-    });
+    this.userService.current()
+      .switchMap((user) => {
+        const id = user.primary_organization.location.properties.api_city_id;
+        return this.cityService.get(id);
+      })
+      .subscribe((apiCity) => {
+        this.apiCity = apiCity;
+      });
     this.setDisabled(this.risk);
 
     this.sessionSubscription = this.session.data.subscribe(risk => {