+++ src/angular/planit/src/app/shared/models/location.model.ts
@@ -10,8 +10,6 @@ export class Location {
   type: string;
   geometry: Point;
   properties: LocationProperties;
-  id: string;
-  dataset: Array<string>;
 
   constructor(object: Object) {
     Object.assign(this, object);